repalce cra to vite (custom payment flow example)
