+++ custom-payment-flow/client/react-cra/index.html
@@ -8,5 +8,6 @@
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <main id="root"></main>
+    <script type="module" src="src/index.jsx"></script>
   </body>
 </html>
