+++ custom-payment-flow/client/react-cra/src/ApplePay.jsx
@@ -33,7 +33,7 @@ const ApplePay = () => {
 
     pr.on('paymentmethod', async (e) => {
       const {error: backendError, clientSecret} = await fetch(
-        '/create-payment-intent',
+        '/api/create-payment-intent',
         {
           method: 'POST',
           headers: {
