+++ custom-payment-flow/client/react-cra/src/BecsDebit.jsx
@@ -30,7 +30,7 @@ const BecsDebitForm = () => {
     }
 
     const {error: backendError, clientSecret} = await fetch(
-      '/create-payment-intent',
+      '/api/create-payment-intent',
       {
         method: 'POST',
         headers: {
