+++ custom-payment-flow/client/react-cra/src/Boleto.jsx
@@ -29,7 +29,7 @@ const Boleto = () => {
     }
 
     const {error: backendError, clientSecret} = await fetch(
-      '/create-payment-intent',
+      '/api/create-payment-intent',
       {
         method: 'POST',
         headers: {
