+++ custom-payment-flow/client/react-cra/src/Eps.jsx
@@ -1,15 +1,12 @@
-import React, {useEffect} from 'react';
+import React, {useEffect, useState} from 'react';
 import {useLocation} from 'react-router-dom';
-import {
-  IdealBankElement,
-  useStripe,
-  useElements,
-} from '@stripe/react-stripe-js';
+import {EpsBankElement, useStripe, useElements} from '@stripe/react-stripe-js';
 import StatusMessages, {useMessages} from './StatusMessages';
 
-const IdealForm = () => {
+const EpsForm = () => {
   const stripe = useStripe();
   const elements = useElements();
+  const [name, setName] = useState('Jenny Rosen');
   const [messages, addMessage] = useMessages();
 
   const handleSubmit = async (e) => {
@@ -24,19 +21,16 @@ const IdealForm = () => {
       return;
     }
 
-    const {error: backendError, clientSecret} = await fetch(
-      '/create-payment-intent',
-      {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          paymentMethodType: 'ideal',
-          currency: 'eur',
-        }),
-      }
-    ).then((r) => r.json());
+    const {error: backendError, clientSecret} = await fetch('/api/create-payment-intent', {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({
+        paymentMethodType: 'eps',
+        currency: 'eur',
+      }),
+    }).then((r) => r.json());
 
     if (backendError) {
       addMessage(backendError.message);
@@ -45,18 +39,18 @@ const IdealForm = () => {
 
     addMessage('Client secret returned');
 
-    const {
-      error: stripeError,
-      paymentIntent,
-    } = await stripe.confirmIdealPayment(clientSecret, {
-      payment_method: {
-        ideal: elements.getElement(IdealBankElement),
-        billing_details: {
-          name: 'Jenny Rosen',
+    const {error: stripeError, paymentIntent} = await stripe.confirmEpsPayment(
+      clientSecret,
+      {
+        payment_method: {
+          eps: elements.getElement(EpsBankElement),
+          billing_details: {
+            name,
+          },
         },
-      },
-      return_url: `${window.location.origin}/ideal?return=true`,
-    });
+        return_url: `${window.location.origin}/eps?return=true`,
+      }
+    );
 
     if (stripeError) {
       // Show error to your customer (e.g., insufficient funds)
@@ -74,21 +68,28 @@ const IdealForm = () => {
 
   return (
     <>
-      <h1>iDEAL</h1>
-      <form id="payment-form" onSubmit={handleSubmit}>
-        <label htmlFor="ideal-bank-element">iDEAL Bank</label>
-        <IdealBankElement id="ideal-bank-element" />
+      <h1>EPS</h1>
 
+      <form id="payment-form" onSubmit={handleSubmit}>
+        <label htmlFor="name">Name</label>
+        <input
+          id="name"
+          value={name}
+          onChange={(e) => setName(e.target.value)}
+          required
+        />
+
+        <EpsBankElement />
         <button type="submit">Pay</button>
       </form>
+
       <StatusMessages messages={messages} />
     </>
   );
 };
 
-// Component for displaying results after returning from
-// iDEAL redirect flow.
-const IdealReturn = () => {
+// Component for displaying results after returning from the redirect flow.
+const EpsReturn = () => {
   const stripe = useStripe();
   const [messages, addMessage] = useMessages();
 
@@ -101,11 +102,11 @@ const IdealReturn = () => {
     }
     const fetchPaymentIntent = async () => {
       const {
-        error,
+        error: stripeError,
         paymentIntent,
       } = await stripe.retrievePaymentIntent(clientSecret);
-      if (error) {
-        addMessage(error.message);
+      if (stripeError) {
+        addMessage(stripeError.message);
       }
       addMessage(`Payment ${paymentIntent.status}: ${paymentIntent.id}`);
     };
@@ -114,19 +115,19 @@ const IdealReturn = () => {
 
   return (
     <>
-      <h1>Ideal Return</h1>
+      <h1>EPS Return</h1>
       <StatusMessages messages={messages} />
     </>
   );
 };
 
-const Ideal = () => {
+const Eps = () => {
   const query = new URLSearchParams(useLocation().search);
   if (query.get('return')) {
-    return <IdealReturn />;
+    return <EpsReturn />;
   } else {
-    return <IdealForm />;
+    return <EpsForm />;
   }
 };
 
-export default Ideal;
+export default Eps;
