+++ custom-payment-flow/client/react-cra/src/Giropay.jsx
@@ -1,11 +1,12 @@
-import React, {useEffect} from 'react';
+import React, {useEffect, useState} from 'react';
 import {useLocation} from 'react-router-dom';
-import {FpxBankElement, useStripe, useElements} from '@stripe/react-stripe-js';
+import {useStripe, useElements} from '@stripe/react-stripe-js';
 import StatusMessages, {useMessages} from './StatusMessages';
 
-const FpxForm = () => {
+const GiropayForm = () => {
   const stripe = useStripe();
   const elements = useElements();
+  const [name, setName] = useState('Jenny Rosen');
   const [messages, addMessage] = useMessages();
 
   const handleSubmit = async (e) => {
@@ -20,16 +21,16 @@ const FpxForm = () => {
       return;
     }
 
-    let {error: backendError, clientSecret} = await fetch(
-      '/create-payment-intent',
+    const {error: backendError, clientSecret} = await fetch(
+      '/api/create-payment-intent',
       {
         method: 'POST',
         headers: {
           'Content-Type': 'application/json',
         },
         body: JSON.stringify({
-          paymentMethodType: 'fpx',
-          currency: 'myr',
+          paymentMethodType: 'giropay',
+          currency: 'eur',
         }),
       }
     ).then((r) => r.json());
@@ -41,15 +42,17 @@ const FpxForm = () => {
 
     addMessage('Client secret returned');
 
-    let {error: stripeError, paymentIntent} = await stripe.confirmFpxPayment(
-      clientSecret,
-      {
-        payment_method: {
-          fpx: elements.getElement(FpxBankElement),
+    const {
+      error: stripeError,
+      paymentIntent,
+    } = await stripe.confirmGiropayPayment(clientSecret, {
+      payment_method: {
+        billing_details: {
+          name,
         },
-        return_url: `${window.location.origin}/fpx?return=true`,
-      }
-    );
+      },
+      return_url: `${window.location.origin}/giropay?return=true`,
+    });
 
     if (stripeError) {
       // Show error to your customer (e.g., insufficient funds)
@@ -67,10 +70,17 @@ const FpxForm = () => {
 
   return (
     <>
-      <h1>FPX</h1>
+      <h1>Giropay</h1>
 
       <form id="payment-form" onSubmit={handleSubmit}>
-        <FpxBankElement options={{accountHolderType: 'individual'}} />
+        <label htmlFor="name">Name</label>
+        <input
+          id="name"
+          value={name}
+          onChange={(e) => setName(e.target.value)}
+          required
+        />
+
         <button type="submit">Pay</button>
       </form>
 
@@ -81,10 +91,11 @@ const FpxForm = () => {
 
 // Component for displaying results after returning from
 // bancontact redirect flow.
-const FpxReturn = () => {
+const GiropayReturn = () => {
   const stripe = useStripe();
   const [messages, addMessage] = useMessages();
 
+  // Extract the client secret from the query string params.
   const query = new URLSearchParams(useLocation().search);
   const clientSecret = query.get('payment_intent_client_secret');
 
@@ -93,9 +104,10 @@ const FpxReturn = () => {
       return;
     }
     const fetchPaymentIntent = async () => {
-      const {error, paymentIntent} = await stripe.retrievePaymentIntent(
-        clientSecret
-      );
+      const {
+        error,
+        paymentIntent,
+      } = await stripe.retrievePaymentIntent(clientSecret);
       if (error) {
         addMessage(error.message);
       }
@@ -106,19 +118,19 @@ const FpxReturn = () => {
 
   return (
     <>
-      <h1>FPX Return</h1>
+      <h1>Giropay Return</h1>
       <StatusMessages messages={messages} />
     </>
   );
 };
 
-const Fpx = () => {
+const Giropay = () => {
   const query = new URLSearchParams(useLocation().search);
   if (query.get('return')) {
-    return <FpxReturn />;
+    return <GiropayReturn />;
   } else {
-    return <FpxForm />;
+    return <GiropayForm />;
   }
 };
 
-export default Fpx;
+export default Giropay;
