+++ custom-payment-flow/client/react-cra/src/GrabPay.jsx
@@ -22,7 +22,7 @@ const GrabPayForm = () => {
     }
 
     const {error: backendError, clientSecret} = await fetch(
-      '/create-payment-intent',
+      '/api/create-payment-intent',
       {
         method: 'POST',
         headers: {
