+++ custom-payment-flow/client/react-cra/src/Konbini.jsx
@@ -23,7 +23,7 @@ const KonbiniForm = () => {
     }
 
     const {error: backendError, clientSecret} = await fetch(
-      '/create-payment-intent',
+      '/api/create-payment-intent',
       {
         method: 'POST',
         headers: {
