+++ custom-payment-flow/client/react-cra/src/Oxxo.jsx
@@ -1,12 +1,12 @@
-import React, {useEffect, useState} from 'react';
-import {useLocation} from 'react-router-dom';
+import React, {useState} from 'react';
 import {useStripe, useElements} from '@stripe/react-stripe-js';
 import StatusMessages, {useMessages} from './StatusMessages';
 
-const GiropayForm = () => {
+const OxxoForm = () => {
   const stripe = useStripe();
   const elements = useElements();
   const [name, setName] = useState('Jenny Rosen');
+  const [email, setEmail] = useState('jr.succeed_immediately@example.com');
   const [messages, addMessage] = useMessages();
 
   const handleSubmit = async (e) => {
@@ -22,15 +22,15 @@ const GiropayForm = () => {
     }
 
     const {error: backendError, clientSecret} = await fetch(
-      '/create-payment-intent',
+      '/api/create-payment-intent',
       {
         method: 'POST',
         headers: {
           'Content-Type': 'application/json',
         },
         body: JSON.stringify({
-          paymentMethodType: 'giropay',
-          currency: 'eur',
+          paymentMethodType: 'oxxo',
+          currency: 'mxn',
         }),
       }
     ).then((r) => r.json());
@@ -42,17 +42,17 @@ const GiropayForm = () => {
 
     addMessage('Client secret returned');
 
-    const {
-      error: stripeError,
-      paymentIntent,
-    } = await stripe.confirmGiropayPayment(clientSecret, {
-      payment_method: {
-        billing_details: {
-          name,
+    const {error: stripeError, paymentIntent} = await stripe.confirmOxxoPayment(
+      clientSecret,
+      {
+        payment_method: {
+          billing_details: {
+            name,
+            email,
+          },
         },
-      },
-      return_url: `${window.location.origin}/giropay?return=true`,
-    });
+      }
+    );
 
     if (stripeError) {
       // Show error to your customer (e.g., insufficient funds)
@@ -66,11 +66,28 @@ const GiropayForm = () => {
     // payment_intent.succeeded event that handles any business critical
     // post-payment actions.
     addMessage(`Payment ${paymentIntent.status}: ${paymentIntent.id}`);
+
+    // When passing {any_prefix}succeed_immediately@{any_suffix}
+    // as the email address in the billing details, the payment
+    // intent will succeed after 3 seconds. We set this timeout
+    // to refetch the payment intent.
+    const i = setInterval(async () => {
+      const {error: e, paymentIntent} = await stripe.retrievePaymentIntent(
+        clientSecret
+      );
+      addMessage(`Payment ${paymentIntent.status}: ${paymentIntent.id}`);
+      if (paymentIntent.status === 'succeeded') {
+        clearInterval(i);
+      }
+      if (e) {
+        addMessage(e.message);
+      }
+    }, 500);
   };
 
   return (
     <>
-      <h1>Giropay</h1>
+      <h1>OXXO</h1>
 
       <form id="payment-form" onSubmit={handleSubmit}>
         <label htmlFor="name">Name</label>
@@ -81,56 +98,23 @@ const GiropayForm = () => {
           required
         />
 
+        <label htmlFor="email">Email</label>
+        <input
+          id="email"
+          type="email"
+          value={email}
+          onChange={(e) => setEmail(e.target.value)}
+          required
+        />
+
         <button type="submit">Pay</button>
       </form>
 
       <StatusMessages messages={messages} />
+
+      <p> <a href="https://youtu.be/zmNMMBbYFf0" target="_blank">Watch a demo walkthrough</a> </p>
     </>
   );
 };
 
-// Component for displaying results after returning from
-// bancontact redirect flow.
-const GiropayReturn = () => {
-  const stripe = useStripe();
-  const [messages, addMessage] = useMessages();
-
-  // Extract the client secret from the query string params.
-  const query = new URLSearchParams(useLocation().search);
-  const clientSecret = query.get('payment_intent_client_secret');
-
-  useEffect(() => {
-    if (!stripe) {
-      return;
-    }
-    const fetchPaymentIntent = async () => {
-      const {
-        error,
-        paymentIntent,
-      } = await stripe.retrievePaymentIntent(clientSecret);
-      if (error) {
-        addMessage(error.message);
-      }
-      addMessage(`Payment ${paymentIntent.status}: ${paymentIntent.id}`);
-    };
-    fetchPaymentIntent();
-  }, [clientSecret, stripe, addMessage]);
-
-  return (
-    <>
-      <h1>Giropay Return</h1>
-      <StatusMessages messages={messages} />
-    </>
-  );
-};
-
-const Giropay = () => {
-  const query = new URLSearchParams(useLocation().search);
-  if (query.get('return')) {
-    return <GiropayReturn />;
-  } else {
-    return <GiropayForm />;
-  }
-};
-
-export default Giropay;
+export default OxxoForm;
