+++ custom-payment-flow/client/react-cra/src/SepaDebit.jsx
@@ -22,7 +22,7 @@ const SepaDebitForm = () => {
     }
 
     const {error: backendError, clientSecret} = await fetch(
-      '/create-payment-intent',
+      '/api/create-payment-intent',
       {
         method: 'POST',
         headers: {
