+++ custom-payment-flow/client/react-cra/src/Sofort.jsx
@@ -1,27 +1,13 @@
 import React, {useEffect, useState} from 'react';
 import {useLocation} from 'react-router-dom';
-import {useStripe} from '@stripe/react-stripe-js';
+import {useStripe, useElements} from '@stripe/react-stripe-js';
 import StatusMessages, {useMessages} from './StatusMessages';
 
-const COUNTRY_CURRENCY = {
-  AT: 'EUR',
-  BE: 'EUR',
-  DK: 'DKK',
-  FI: 'EUR',
-  DE: 'EUR',
-  IT: 'EUR',
-  NL: 'EUR',
-  NO: 'NOK',
-  ES: 'EUR',
-  SE: 'SEK',
-  UK: 'GBP',
-  US: 'USD',
-}
-
-const KlarnaForm = () => {
+const SofortForm = () => {
   const stripe = useStripe();
-  const [email, setEmail] = useState('user-us@example.com');
-  const [country, setCountry] = useState('US');
+  const elements = useElements();
+  const [name, setName] = useState('Jenny Rosen');
+  const [email, setEmail] = useState('jenny.rosen@example.com');
   const [messages, addMessage] = useMessages();
 
   const handleSubmit = async (e) => {
@@ -29,7 +15,7 @@ const KlarnaForm = () => {
     // which would refresh the page.
     e.preventDefault();
 
-    if (!stripe) {
+    if (!stripe || !elements) {
       // Stripe.js has not yet loaded.
       // Make sure to disable form submission until Stripe.js has loaded.
       addMessage('Stripe.js has not yet loaded.');
@@ -37,15 +23,15 @@ const KlarnaForm = () => {
     }
 
     const {error: backendError, clientSecret} = await fetch(
-      '/create-payment-intent',
+      '/api/create-payment-intent',
       {
         method: 'POST',
         headers: {
           'Content-Type': 'application/json',
         },
         body: JSON.stringify({
-          paymentMethodType: 'klarna',
-          currency: COUNTRY_CURRENCY[country],
+          paymentMethodType: 'sofort',
+          currency: 'eur',
         }),
       }
     ).then((r) => r.json());
@@ -57,20 +43,21 @@ const KlarnaForm = () => {
 
     addMessage('Client secret returned');
 
-    const {error: stripeError, paymentIntent} = await stripe.confirmKlarnaPayment(
-      clientSecret,
-      {
-        payment_method: {
-          billing_details: {
-            email,
-            address: {
-              country
-            }
-          },
+    const {
+      error: stripeError,
+      paymentIntent,
+    } = await stripe.confirmSofortPayment(clientSecret, {
+      payment_method: {
+        sofort: {
+          country: 'DE',
         },
-        return_url: `${window.location.origin}/klarna?return=true`,
-      }
-    );
+        billing_details: {
+          name,
+          email,
+        },
+      },
+      return_url: `${window.location.origin}/sofort?return=true`,
+    });
 
     if (stripeError) {
       // Show error to your customer (e.g., insufficient funds)
@@ -88,44 +75,41 @@ const KlarnaForm = () => {
 
   return (
     <>
-      <h1>Klarna</h1>
-
-      <table>
-        <thead><tr><th><div ><div></div></div></th><th><div><div><div>Approved</div></div></div></th><th><div><div><div>Denied</div></div></div></th></tr></thead>
-        <tbody>
-          <tr><td>Email</td><td>user-us@example.com</td><td>user-us+denied@example.com</td></tr>
-          <tr><td>Street</td><td>Lombard St 10</td><td>Lombard St 10</td></tr>
-          <tr><td>Apartment</td><td>Apt 214</td><td>Apt 214</td></tr>
-          <tr><td>City</td><td>Beverly Hills</td><td>Beverly Hills</td></tr>
-          <tr><td>State</td><td>CA</td><td>CA</td></tr>
-          <tr><td>Postal code</td><td>90210</td><td>90210</td></tr>
-          <tr><td>Phone</td><td>3106143666</td><td>3106143888</td></tr>
-        </tbody>
-      </table>
+      <h1>Sofort</h1>
 
       <form id="payment-form" onSubmit={handleSubmit}>
-        <label htmlFor="email">Email</label>
-        <input type="email" id="email" placeholder="jenny.rosen@example.com" value={email} onChange={(e) => setEmail(e.target.value)} required />
+        <label htmlFor="name">Name</label>
+        <input
+          id="name"
+          value={name}
+          onChange={(e) => setName(e.target.value)}
+          required
+        />
 
-        <label htmlFor="country">
-          Country
-        </label>
-        <select id="country" value={country}>
-          {Object.keys(COUNTRY_CURRENCY).map(country => <option key={country}>{country}</option>)}
-        </select>
+        <label htmlFor="email">Email</label>
+        <input
+          id="email"
+          type="email"
+          value={email}
+          onChange={(e) => setEmail(e.target.value)}
+          required
+        />
 
         <button type="submit">Pay</button>
       </form>
+
       <StatusMessages messages={messages} />
     </>
   );
 };
 
-// Component for displaying results after returning from the redirect flow.
-const KlarnaReturn = () => {
+// Component for displaying results after returning from
+// bancontact redirect flow.
+const SofortReturn = () => {
   const stripe = useStripe();
   const [messages, addMessage] = useMessages();
 
+  // Extract the client secret from the query string params.
   const query = new URLSearchParams(useLocation().search);
   const clientSecret = query.get('payment_intent_client_secret');
 
@@ -148,19 +132,19 @@ const KlarnaReturn = () => {
 
   return (
     <>
-      <h1>Klarna Return</h1>
+      <h1>Sofort Return</h1>
       <StatusMessages messages={messages} />
     </>
   );
 };
 
-const Klarna = () => {
+const Sofort = () => {
   const query = new URLSearchParams(useLocation().search);
   if (query.get('return')) {
-    return <KlarnaReturn />;
+    return <SofortReturn />;
   } else {
-    return <KlarnaForm />;
+    return <SofortForm />;
   }
-}
+};
 
-export default Klarna;
+export default Sofort;
