+++ custom-payment-flow/client/react-cra/src/UsBankAccountDebit.jsx
@@ -28,7 +28,7 @@ const UsBankAccountDebitForm = () => {
     }
 
     const {error: backendError, clientSecret} = await fetch(
-      '/create-payment-intent',
+      '/api/create-payment-intent',
       {
         method: 'POST',
         headers: {
