+++ custom-payment-flow/client/react-cra/vite.config.mjs
@@ -1,0 +1,21 @@
+import react from "@vitejs/plugin-react"
+import { defineConfig } from "vite"
+
+export default defineConfig({
+  plugins: [
+    react(),
+  ],
+  server: {
+    port: 3000,
+    proxy: {
+        '/api': {
+          target: 'http://localhost:4242',
+          changeOrigin: true,
+          rewrite: (path) => path.replace(/^\/api/, '')
+        }
+    }
+  },
+  build: {
+    outDir: "build",
+  },
+})
